sprint_5/pyproject.toml
[tool.poetry]
name = "YP"
version = "0.1.0"
description = ""
authors = ["Galovach Andrey <a.galovach@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
redis = "^4.3.4"
flake8-bugbear = "^22.9.23"
Django = "^4.1.1"
django-split-settings = "^1.1.0"
django-debug-toolbar = "^3.7.0"
django-extensions = "^3.2.1"
django-cors-headers = "^3.13.0"
uWSGI = "^2.0.20"
orjson = "^3.8.0"
elasticsearch = {version = "7.9.1", extras = ["async"]}
aioredis = "^2.0.1"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
elasticsearch-dsl = "^7.4.0"
gunicorn = "^20.1.0"
aiohttp = "^3.8.3"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]
flake8-bugbear = "^22.9.23"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.3.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-variables-names = "^0.0.5"
pre-commit = "^2.19.0"
dlint = "^0.13.0"
black = "^22.6.0"
isort = "^5.10.1"
yesqa = "^1.4.0"
flake8 = "^5.0.4"
pytest = "^7.1.2"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
